Adicionado os endpoints da entidade Cidade e o endpoint Path para a entidade Restaurante

+++ algafood-api/src/main/java/com/algaworks/algafood_api/domain/service/CadastroEstadoService.java
@@ -22,22 +22,24 @@ public class CadastroEstadoService {
     }
 
     public Estado atualizar(Long estadoId, Estado estadoAlterar){
-        try{
+
             Estado estadoEncontrado = estadoRepository.buscar(estadoId);
+            if(estadoEncontrado == null){
+                throw new EstadoNaoEncontradoException(String.format("Não existe cadastro de estado com o id %d", estadoId));
+            }
             BeanUtils.copyProperties(estadoAlterar, estadoEncontrado, "id");
             estadoEncontrado = estadoRepository.salvar(estadoEncontrado);
             return estadoEncontrado;
-        } catch (EmptyResultDataAccessException e) {
-            throw new EstadoNaoEncontradoException(String.format("Não existe cadastro de estado com o id %d", estadoId));
-        }
+
     }
 
     public void remover(Long estadoId){
         try{
             Estado estadoEncontrado = estadoRepository.buscar(estadoId);
+            if(estadoEncontrado == null){
+                throw new EstadoNaoEncontradoException(String.format("Não existe cadastro de estado com o id %d", estadoId));
+            }
             estadoRepository.remover(estadoEncontrado);
-        } catch (EmptyResultDataAccessException e){
-            throw new EstadoNaoEncontradoException(String.format("Não foi encontrado nenhum estado com o id %d", estadoId));
         } catch (DataIntegrityViolationException e) {
             throw new EntidadeEmUsoException(String.format("Entidade com id %d não pode ser removida, pois está em uso", estadoId));
         }
